Flatten device settings into the main settings page
Merge the separate Device Settings (glasses) page into the main settings
screen so it's a single page. Tapping the glasses tile on the home screen
now opens the flattened main settings instead of a dedicated device page.

- Extract the device-settings body into a shared DeviceSettingsSection
  component, preserving every capability/device conditional verbatim.
- Render it inline on the main settings page (showAdvancedGroup=false);
  fold Device Information into the page-level Advanced Settings group and
  drop the duplicate Microphone row (already under App Settings).
- Keep /miniapps/settings/glasses resolvable as a thin wrapper around the
  same component for the offline miniapp host and deep links.
- Repoint the home tile and the wifi-scan back-route to the main page.
