+++ mobile/src/app/miniapps/settings/main.tsx
@@ -3,6 +3,7 @@ import {ScrollView} from "react-native-gesture-handler"
 
 import {VersionInfo} from "@/components/dev/VersionInfo"
 import {Icon, Screen} from "@/components/ignite"
+import {DeviceSettingsSection, useHasDeviceInfo} from "@/components/settings/DeviceSettingsSection"
 import {Group} from "@/components/ui/Group"
 import {RouteButton} from "@/components/ui/RouteButton"
 import {Spacer} from "@/components/ui/Spacer"
@@ -18,7 +19,7 @@ export default function MainSettingsPage() {
   const {push} = useNavigationStore.getState()
   const [debugMode] = useSetting(SETTINGS.debug_mode.key)
   const [superMode] = useSetting(SETTINGS.super_mode.key)
-  const [defaultWearable] = useSetting(SETTINGS.default_wearable.key)
+  const hasDeviceInfo = useHasDeviceInfo()
   const viewShotRef = useRef<View>(null)
 
   useRegisterCapsule({
@@ -29,84 +30,85 @@ export default function MainSettingsPage() {
   })
 
   return (
-      <Screen preset="fixed" safeAreaEdges={["top"]} ref={viewShotRef} className="px-0">
-        <ScrollView className="pt-8 px-6" contentInsetAdjustmentBehavior="automatic">
-          <View style={{flex: 1, gap: theme.spacing.s6}}>
-            <Group title={translate("account:accountSettings")}>
-              <RouteButton
-                icon={<Icon name="circle-user" size={24} color={theme.colors.secondary_foreground} />}
-                label={translate("settings:profileSettings")}
-                onPress={() => push("/miniapps/settings/profile")}
-              />
-              <RouteButton
-                icon={<Icon name="message-2-star" size={24} color={theme.colors.secondary_foreground} />}
-                label={translate("settings:feedback")}
-                onPress={() => push("/miniapps/settings/feedback")}
-              />
-            </Group>
+    <Screen preset="fixed" safeAreaEdges={["top"]} ref={viewShotRef} className="px-0">
+      <ScrollView className="pt-8 px-6" contentInsetAdjustmentBehavior="automatic">
+        <View style={{flex: 1, gap: theme.spacing.s6}}>
+          <Group title={translate("account:accountSettings")}>
+            <RouteButton
+              icon={<Icon name="circle-user" size={24} color={theme.colors.secondary_foreground} />}
+              label={translate("settings:profileSettings")}
+              onPress={() => push("/miniapps/settings/profile")}
+            />
+            <RouteButton
+              icon={<Icon name="message-2-star" size={24} color={theme.colors.secondary_foreground} />}
+              label={translate("settings:feedback")}
+              onPress={() => push("/miniapps/settings/feedback")}
+            />
+          </Group>
 
-            {defaultWearable && (
-              <Group title={translate("account:deviceSettings")}>
-                <RouteButton
-                  icon={<Icon name="glasses" color={theme.colors.secondary_foreground} size={24} />}
-                  label={defaultWearable}
-                  onPress={() => push("/miniapps/settings/glasses")}
-                />
-              </Group>
+          {/* Device/glasses settings, flattened inline (previously a separate page).
+              Device Information lives in the page-level Advanced Settings group below. */}
+          <DeviceSettingsSection showAdvancedGroup={false} />
+
+          <Group title={translate("account:appSettings")}>
+            {superMode && (
+              <RouteButton
+                icon={<Icon name="sun" size={24} color={theme.colors.secondary_foreground} />}
+                label={translate("settings:appAppearance")}
+                onPress={() => push("/miniapps/settings/appearance")}
+              />
             )}
+            {(Platform.OS === "android" || superMode) && (
+              <RouteButton
+                icon={<Icon name="bell" size={24} color={theme.colors.secondary_foreground} />}
+                label={translate("settings:notificationsSettings")}
+                onPress={() => push("/miniapps/settings/notifications")}
+              />
+            )}
+            <RouteButton
+              icon={<Icon name="microphone" size={24} color={theme.colors.secondary_foreground} />}
+              label={translate("deviceSettings:microphone")}
+              onPress={() => push("/miniapps/settings/microphone")}
+            />
+            <RouteButton
+              icon={<Icon name="volume" size={24} color={theme.colors.secondary_foreground} />}
+              label={translate("settings:speechSettings")}
+              onPress={() => push("/miniapps/settings/speech")}
+            />
+            <RouteButton
+              icon={<Icon name="shield-lock" size={24} color={theme.colors.secondary_foreground} />}
+              label={translate("settings:privacySettings")}
+              onPress={() => push("/miniapps/settings/privacy")}
+            />
+          </Group>
 
-            <Group title={translate("account:appSettings")}>
-              {superMode && (
-                <RouteButton
-                  icon={<Icon name="sun" size={24} color={theme.colors.secondary_foreground} />}
-                  label={translate("settings:appAppearance")}
-                  onPress={() => push("/miniapps/settings/appearance")}
-                />
-              )}
-              {(Platform.OS === "android" || superMode) && (
-                <RouteButton
-                  icon={<Icon name="bell" size={24} color={theme.colors.secondary_foreground} />}
-                  label={translate("settings:notificationsSettings")}
-                  onPress={() => push("/miniapps/settings/notifications")}
-                />
-              )}
+          <Group title={translate("deviceSettings:advancedSettings")}>
+            {hasDeviceInfo && (
               <RouteButton
-                icon={<Icon name="microphone" size={24} color={theme.colors.secondary_foreground} />}
-                label={translate("deviceSettings:microphone")}
-                onPress={() => push("/miniapps/settings/microphone")}
+                icon={<Icon name="device-ipad" size={24} color={theme.colors.secondary_foreground} />}
+                label={translate("deviceSettings:deviceInformation")}
+                onPress={() => push("/miniapps/settings/device-info")}
               />
-              <RouteButton
-                icon={<Icon name="volume" size={24} color={theme.colors.secondary_foreground} />}
-                label={translate("settings:speechSettings")}
-                onPress={() => push("/miniapps/settings/speech")}
-              />
-              <RouteButton
-                icon={<Icon name="shield-lock" size={24} color={theme.colors.secondary_foreground} />}
-                label={translate("settings:privacySettings")}
-                onPress={() => push("/miniapps/settings/privacy")}
-              />
-            </Group>
-
-            <Group title={translate("deviceSettings:advancedSettings")}>
-              {debugMode && (
-                <RouteButton
-                  icon={<Icon name="user-code" size={24} color={theme.colors.secondary_foreground} />}
-                  label={translate("settings:debugSettings")}
-                  onPress={() => push("/miniapps/settings/debug")}
-                  onLongPress={() => superMode && push("/miniapps/settings/super")}
-                />
-              )}
+            )}
+            {debugMode && (
               <RouteButton
                 icon={<Icon name="user-code" size={24} color={theme.colors.secondary_foreground} />}
-                label={translate("settings:miniappDeveloperSettings")}
-                onPress={() => push("/miniapps/settings/miniapp-dev")}
+                label={translate("settings:debugSettings")}
+                onPress={() => push("/miniapps/settings/debug")}
+                onLongPress={() => superMode && push("/miniapps/settings/super")}
               />
-            </Group>
-          </View>
+            )}
+            <RouteButton
+              icon={<Icon name="user-code" size={24} color={theme.colors.secondary_foreground} />}
+              label={translate("settings:miniappDeveloperSettings")}
+              onPress={() => push("/miniapps/settings/miniapp-dev")}
+            />
+          </Group>
+        </View>
 
-          <VersionInfo />
-          <Spacer height={theme.spacing.s10} />
-        </ScrollView>
-      </Screen>
+        <VersionInfo />
+        <Spacer height={theme.spacing.s10} />
+      </ScrollView>
+    </Screen>
   )
 }
